udemy-deploy-ml-models-with-streamlit/pyproject.toml
[project]
name = "udemy-deploy-ml-models-with-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit",
    "pandas",
    "matplotlib",
    "numpy",
    "scipy",
    "scikit-learn>=1.7.2",
    "streamlit_authenticator",
    "PyYAML",
    "supabase>=2.22.0",
    "statsforecast>=2.0.1",
]
